+++ src/test/java/me/mixss/dynatracebackendtask/services/MaxDiffBuyAndSellRateServiceTest.java
@@ -33,7 +33,7 @@ class MaxDiffBuyAndSellRateServiceTest {
         JsonNode apiResult = mapper.readTree("{\"table\":\"C\",\"currency\":\"dolar amerykański\",\"code\":\"USD\",\"rates\":[{\"no\":\"074/C/NBP/2023\",\"effectiveDate\":\"2023-04-17\",\"bid\":4.1626,\"ask\":4.2466},{\"no\":\"075/C/NBP/2023\",\"effectiveDate\":\"2023-04-18\",\"bid\":4.1919,\"ask\":4.2765},{\"no\":\"076/C/NBP/2023\",\"effectiveDate\":\"2023-04-19\",\"bid\":4.1769,\"ask\":4.2613},{\"no\":\"077/C/NBP/2023\",\"effectiveDate\":\"2023-04-20\",\"bid\":4.1677,\"ask\":4.2519},{\"no\":\"078/C/NBP/2023\",\"effectiveDate\":\"2023-04-21\",\"bid\":4.1532,\"ask\":4.2372}]}");
         String currencyCode = "usd";
         int numberOfQuotations = 5;
-        float correctResult = 0.0846f;
+        double correctResult = 0.0846;
 
         when(apiClient.getLastQuotationsBuyAndSellRates(currencyCode, numberOfQuotations)).thenReturn(apiResult);
